Drop funopen usage from BSD and Linux
We don't need 3 implementations of for simultaneous streams.
Drop funopen from Linux and *BSD, and use Sun's ESTREAM solution
for all platforms.  This makes the Ada interface much simpler
and hopefully improves the robustness (to avoid the possibility
of fetch working on some platforms and not others).

Also convert to CMake for portability